Move exception formatting logic into its own class.
- Simplifies the notification classes.
- Provides something we can use to format multiple
  expectation failures packaged as a single exception
  for `aggregate_failures`.
