+++ spec/rspec/core/notifications_spec.rb
@@ -1,10 +1,11 @@
 require 'rspec/core/notifications'
-require 'pathname'
 
 RSpec.describe "FailedExampleNotification" do
   include FormatterSupport
 
   let(:example) { new_example }
+  exception_line = __LINE__ + 1
+  let(:exception) { instance_double(Exception, :backtrace => [ "#{__FILE__}:#{exception_line}"], :message => 'Test exception') }
   let(:notification) { ::RSpec::Core::Notifications::FailedExampleNotification.new(example) }
 
   before do
@@ -12,86 +13,16 @@ RSpec.describe "FailedExampleNotification" do
     example.metadata[:absolute_file_path] = __FILE__
   end
 
-  # ported from `base_formatter_spec` should be refactored by final
-  describe "#read_failed_line" do
-    context "when backtrace is a heterogeneous language stack trace" do
-      let(:exception) do
-        instance_double(Exception, :backtrace => [
-          "at Object.prototypeMethod (foo:331:18)",
-          "at Array.forEach (native)",
-          "at a_named_javascript_function (/some/javascript/file.js:39:5)",
-          "/some/line/of/ruby.rb:14"
-        ])
-      end
+  it 'provides a description' do
+    expect(notification.description).to eq(example.full_description)
+  end
 
-      it "is handled gracefully" do
-        expect { notification.send(:read_failed_line) }.not_to raise_error
-      end
-    end
-
-    context "when backtrace will generate a security error" do
-      let(:exception) { instance_double(Exception, :backtrace => [ "#{__FILE__}:#{__LINE__}"]) }
-
-      it "is handled gracefully" do
-        with_safe_set_to_level_that_triggers_security_errors do
-          expect { notification.send(:read_failed_line) }.not_to raise_error
-        end
-      end
-    end
-
-    context "when ruby reports a bogus line number in the stack trace" do
-      let(:exception) { instance_double(Exception, :backtrace => [ "#{__FILE__}:10000000"]) }
-
-      it "reports the filename and that it was unable to find the matching line" do
-        expect(notification.send(:read_failed_line)).to include("Unable to find matching line")
-      end
-    end
-
-    context "when ruby reports a file that does not exist" do
-      let(:file) { "#{__FILE__}/blah.rb" }
-      let(:exception) { instance_double(Exception, :backtrace => [ "#{file}:1"]) }
-
-      it "reports the filename and that it was unable to find the matching line" do
-        example.metadata[:absolute_file_path] = file
-        expect(notification.send(:read_failed_line)).to include("Unable to find #{file} to read failed line")
-      end
-    end
-
-    context "when the stacktrace includes relative paths (which can happen when using `rspec/autorun` and running files through `ruby`)" do
-      let(:relative_file) { Pathname(__FILE__).relative_path_from(Pathname(Dir.pwd)) }
-      line = __LINE__
-      let(:exception) { instance_double(Exception, :backtrace => ["#{relative_file}:#{line}"]) }
-
-      it 'still finds the backtrace line' do
-        expect(notification.send(:read_failed_line)).to include("line = __LINE__")
-      end
-    end
-
-    context "when String alias to_int to_i" do
-      before do
-        String.class_exec do
-          alias :to_int :to_i
-        end
-      end
-
-      after do
-        String.class_exec do
-          undef to_int
-        end
-      end
-
-      let(:exception) { instance_double(Exception, :backtrace => [ "#{__FILE__}:#{__LINE__}"]) }
-
-      it "doesn't hang when file exists" do
-        expect(notification.send(:read_failed_line).strip).to eql(
-          %Q[let(:exception) { instance_double(Exception, :backtrace => [ "\#{__FILE__}:\#{__LINE__}"]) }])
-      end
-
-    end
+  it 'provides `colorized_formatted_backtrace`, which formats the backtrace and colorizes it' do
+    allow(RSpec.configuration).to receive(:color_enabled?).and_return(true)
+    expect(notification.colorized_formatted_backtrace).to eq(["\e[36m# #{RSpec::Core::Metadata.relative_path(__FILE__)}:#{exception_line}\e[0m"])
   end
 
   describe '#message_lines' do
-    let(:exception) { instance_double(Exception, :backtrace => [ "#{__FILE__}:#{__LINE__}"], :message => 'Test exception') }
     let(:example_group) { class_double(RSpec::Core::ExampleGroup, :metadata => {}, :parent_groups => [], :location => "#{__FILE__}:#{__LINE__}") }
 
     before do
